+++ bravado_types/__init__.py
@@ -3,23 +3,24 @@ from typing import Union
 from bravado.client import SwaggerClient
 from bravado_core.spec import Spec
 
+from bravado_types.config import Config
 from bravado_types.extract import get_spec_info
 from bravado_types.metadata import get_metadata
-from bravado_types.render import RenderConfig, render
+from bravado_types.render import render
 
 
 def generate_module(client_or_spec: Union[SwaggerClient, Spec],
-                    render_config: RenderConfig) -> None:
+                    config: Config) -> None:
     """
     Convenience function for extracting spec info and rendering files.
 
     :param client_or_spec: Swagger client or spec.
-    :param render_config: Rendering configuration.
+    :param config: Configuration parameters.
     """
     if isinstance(client_or_spec, SwaggerClient):
         spec = client_or_spec.swagger_spec
     else:
         spec = client_or_spec
     metadata = get_metadata(spec)
-    spec_info = get_spec_info(spec)
-    render(metadata, spec_info, render_config)
+    spec_info = get_spec_info(spec, config)
+    render(metadata, spec_info, config)