+++ CHANGES.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+- Move config class to separate module
+- Add support for custom formats
+- Eagerly generate type strings during type extraction phase
+
 ## 0.3.0
 
 - Specify minimum versions of bravado and bravado-core in setup.py